Remove Redis::ActiveSupport::VERSION
Basically, there is no way to have this constant exist,
cleanly, without shadowing ActiveSupport::VERSION (which
is actually a different type).

This happens due to constant lookups:
http://stem.ps/rails/2015/01/25/ruby-gotcha-toplevel-constant-referenced-by.html

In theory, it would be much better not to invade/use either ActiveSupport
or Redis constants, but have our own constant (like RedisActiveSupport) to stuff
all the modules and constants into, but this would be a major version change.