Support Rails 5 (ActiveSupport 5)
The current version of redis-activesupport works with Rails 5.0.0.beta1
and shows some deprecation warnings.

This commit is meant to upgrade the code of redis-activesupport to remove
those deprecation warnings.

Given the large matrix of tests that run for redis-activesupport,
I will open a PR to check what the backward compatibility looks like.