Store#increment and #decrement must support options
According to the Rails API guide for a Store,
a third argument is accepted for increment()
and decrement(), and the store can elect to
use that argument or ignore it. If the store
is wrapped into another object for delegation,
the method signature of the store should match
the expected signature of the ActiveSupport method.

http://api.rubyonrails.org/classes/ActiveSupport/Cache/Store.html#method-i-increment
http://api.rubyonrails.org/classes/ActiveSupport/Cache/Store.html#method-i-decrement

The fat that this was not the case warranted
patching our cache store wrapper to ensure
the divergent method signature is used instead.