Support regular expressions in delete_matched
Rails uses the type of the argument passed into `expire_fragment` to
determine whether it should be sent to
`ActiveSupport::Cache::Store#delete_matched` or
`ActiveSupport::Cache::Store#delete`. Regular expressions are passed
into `delete_matched`, so remove the error causing this to be impossible
in `RedisStore`, and ensure the regex is being stripped of non-essential
data before it's sent to Redis.

Note that there are no guarantees any of these regex's will work when
they get to Redis, we're just allowing Rails to use the correct method
here.

Fixes #56