Support recyclable cache keys
ActiveSupport 5.2.0 introduces the concept of
[recyclable cache keys](https://github.com/rails/rails/pull/29092),
which prevents a bunch of unnecessary keys being created and then having
to be evicted. This should reduce the memory overhead of keeping a Redis
or Memcache server up in order to support a cache. This caused issues
for us because the `options` are only being written into the cache key,
not the actual entry, which is now required by Rails' cache store
semantics. This should fix issues that people are having with rails 5.2.0.rc1
and redis-activesupport.