Relax runtime dependency on `active_support` for Rails 6
Rails 6 is now in its second beta which means a release candidate could
be on the horizon and then a full release.

In the previous commit I setup an appraisal gemfile for Rails
6.0.0.beta2 and I am fairly confident that the gem still works.

By relaxing the runtime dependency on active_support we can introduce
explicit support for Rails 6.

**Notes**

To be on the safe side I have kept the max allowed version below 7.